Casting fixes.

+++ libtiff/tif_print.c
@@ -73,7 +73,7 @@ TIFFPrintDirectory(TIFF* tif, FILE* fd, long flags)
 	long l, n;
 
 	fprintf(fd, "TIFF Directory at offset 0x%lx\n",
-		(unsigned int)tif->tif_diroff);
+		(unsigned long)tif->tif_diroff);
 	td = &tif->tif_dir;
 	if (TIFFFieldSet(tif,FIELD_SUBFILETYPE)) {
 		fprintf(fd, "  Subfile Type:");
